Read Etherwiki URL from <dir>/.etherwiki/config
Spit out all kinds of warning if it doesn't exist, or isn't valid.

The Daemon class now takes the path to the desired directory as an
argument. If it's not specified, the daemon won't connect to the
Internet.

npm run ethersync will use "./output" as a default directory. You can
specify a different one using

    npm run etherwiki --directory=<dir>