Fix integration tests: Run them in temporary environment
Now integration test suite is creating a temporary directory and
will actually make a connection to the playground.

For the latter I'm quite unsure that this is the best solution.
My thinking was, that "integration test" also could include testing the server component.
Also it was the fastest way to get the tests running, so we can also refine from here :)