If 'fixeol' would insert a \n, immediately send it to the daemon
Do this by setting 'eol' to false if an operation leaves us without a
"real" newline at the end of the content.

Seems to work in manual tests, but in the fuzzer, there's often a
discrepancy of the newlines at the end. Sometimes the fuzzer also
crashes or gets stuck.