Drop editor_is_connected state, as it's a bit redundant, and rather use existence of ot_server itself
The boolean was false, when OT Server was in an invalid state, and true, when it was valid.
Using an Option makes the usage a bit more cumbersome, because we have to unpack it
every time we want to use ot_server, but this way, we can "go wrong",
e.g. set editor_is_connected accidentally to true, even though OTServer is not valid or the other way around.