Replace select! in socket handler with a new task spawn
For a mpsc channel of capacity 1, the fuzzer always seemed to hang. For
bigger capacities, it seemed to *sometimes* hang. I think processing
socket input and output at the same time helps with that. But I haven't
thought a lot about why the problem happens.

This introduces a new TODO: I think we don't know when the client
disconnects now...