Refactor: Create a DocumentActorHandle as an interface to the DocumentActor
This change might be a bit overkill, but I think it does indeed look more
nicely separated afterwards. The main change is, that we split of the responsibility
to set up the DocumentActor (previously called DaemonActor, but I think the rename
gets across the idea better: this is only about the Document resource and has
nothing to do with who connects to it in which way) to the DocumentActorHandle.

The DocumentActorHandle now clearly owns and encapsulates the communication channels
to the DocumentActor. Other actors or parts of the code can send DocMessages, but they
now don't have the responsibility to worry about whether the channel still is alive etc.
Right now, this doesn't change much: Things will panic anyways, just in a different place,
but we now have the callee handle this case.

Also things get a bit less confusing: We don't pass around multiple channels (e.g. the
ping for change channel as well as the "document sender" channel), but only the handle.
So the notion is: Every entity that wants to influence the document needs a DocumentActorHandle.