Refactor: move/restructure the listen_socket loop to the connect module
Similar to TCP.

A note on &Path vs. PathBuf:
After reading https://nick.groenen.me/notes/rust-path-vs-pathbuf/, my
understanding is as follows:
- &Path is the "read only" reference
- PathBuf is the one we want to store.

So, similar to preferring a &str interface over a String interface,
I have now tried to keep the pattern that we only have PathBuf interface
where it's not possible to use &Path.

A weird example of that is the interface of make_editor_connection, which
seems to require PathBuf because of the spawn. It wouldn't work otherwise
(or I would need to explicitly move the &Path, not sure if that's an option).