refactor/protocol refinement: Use LSP's Range format with start/end
This commit replaces all usages of anchor/head by start/end for the
Range that is communicated between editor and daemon.

Motivation is to be closer to the LSP spec in order to make plugin
development smoother. We have observed in the neovim case that many
editor internal LSP methods are useful for our purposes and with this
change we don't have to transform between the diffent data structures.

Closes #27