vscode: When we can't apply a remote edit, send empty edit in response
The reason is this: After applying the remote edit failed, it's not safe
for us to *try* to apply more remote edits for the same editor revision.
They are intended to be applied on top of the previous one, so the
current content we have might not be suitable.

To work around this, send an empty edit to the daemon, and increase our
editor revision. Because of this, we will no longer *try* to apply
remote edits for the old editor revision. The daemon will transform the
left-over edits, and send them again for our new editor revision.