Initial Intellij Plugin Setup
This commit provides the initial structure to create an Intellij plugin
for ethersync. So far the plugin can connect to a remote ethersync
daemon, by starting automatically another daemon, and when the daemon is
ready, it starts a ethersync client which then notifies the plugin about
cursor events, printed in System.out.