Refactorings without Testing
This commit makes the code more idiomatic to the use of Eclipse JSONRPC
and also makes use of some Kotlin coroutines.

Also, the event listeners are initialized in such a way that users don't
have to watch out that files have to be opened after the ethersync
client connection has to established first.