Use <project>/.ethersync/socket as the socket path
This has the advantage that two daemons started on different directories
will never clash - useful for multi-daemon scenarios, and for local
testing.

One downside is that the editor plugins need to do a bit more work:
They need to figure out which directory to start `ethersync client` in.
However, our two plugins already had all the code required to find it,
because they needed to scan for the .ethersync directory anyway.

The --directory flag is now global for all subcommands.