ci: Switch back to rust-cache action
When I tried adding a manual cache action, sometimes, caches couldn't be
saved with the error "Unable to reserve cache with key". We could come
back to it later, and go with this for now.

The downside I see is that, once Cargo.lock changes, the cache won't be
used at all, which slows the CI down a bit. But maybe we don't do that
very often.