Downgrade constant_time_eq to allow MacOS release builds
This is a tricky one: constant_time_eq version 4.0.0 and higher is an edition 2024
crate and requires Rust 1.85 or higher. But from that version on there's a problem
with the MacOS release build, due to a bug down the stack of iroh, described in [1].

As constant_time_eq would be our only edition 2024 crate, we can keep it pinned
to a 2021 edition version and a certain developer using MacOS can keep building
it with Rust 1.84.

[1] https://github.com/n0-computer/iroh/issues/3353