vim: When pressing o on a line, edit should be placed in the new line
Instead of placing the edit after the previous one. This fixes #288.

Strategy: Add a new "fix" branch. Also, when 'eol' is on, pretend there
are empty lines for compute_diff. This leads to prettier diffs overall.
We need that differentiation because otherwise, for buffer content "x",
when pressing o, we have no chance to insert in the next line (which we
may do when 'eol' is on!)