vim: Don't print message that we're disabling 'autoread' on each startup
Move the information into the documentation. I think it belongs there,
it's such a minor issue that we don't need to display it at every
startup.

In addition, in narrow terminal windows, the message forces the user to
press "Enter" to acknowledge it, which is probably more annoying than
being surprised by the changed option.