Remove 'files' map in CRDT document
After reading https://github.com/ethersync/ethersync/issues/69 in
combination with
https://github.com/ethersync/ethersync/pull/152#issuecomment-2311619197
I thought it might actually be worth to remove the hierarchy
and see how far we can get with that.

It simplifies the code and we might have an easier time reasoning
about initialization (hopefully).

The main disadvantage I see is that we can not synchronize
any metadata other than file changes and patches.