vscode: Instead of expecting certain remote edits, expect their result
When a remote edit replaces "x" with "XxX", VS Code will see a
TextDocumentChangeEvent that inserts two X's around the x. Because of
that, it's not possible to directly compare the edits.

Here's a proposed solution: Store the expected resulting content of the
document after the remote edit, then see if a local edit results in that
content.